HSEARCH-3048 Test norms at backend-lucene level
